Added customizable tempUrl duration, changed logic for signature and minimized tempUrl generation
