+++ src/main/java/edu/harvard/iq/dataverse/dataaccess/DataConverter.java
@@ -264,7 +264,7 @@ public class DataConverter {
         }
             
 
-        if (formatConvertedFile.exists()) {
+        if (formatConvertedFile != null && formatConvertedFile.exists()) {
             logger.fine("frmtCnvrtdFile:length=" + formatConvertedFile.length());
         } else {
             logger.warning("Format-converted file was not properly created.");
